% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hisclass}
\alias{hisclass}
\title{Hisclass meta data}
\format{A data frame with 14 rows and 6 variables:
\describe{
  \item{hisclass}{hisclass code}
  \item{hisclass_label}{Description of class}
  \item{manual}{Manual/Non-manual}
  \item{skill_level}{Skill level}
  \item{supervision}{Supervision}
  \item{sector}{Economic sector}
}}
\source{
Leeuwen, M.H.D.V. & Maas, I. (2011). \emph{HISCLASS: a
  historical international social class scheme. Leuven:
  Leuven University Press};  Mandemakers etal. (2013)
  \emph{HSN standardized, HISCO-coded and classified occupational
  titles, release 2013.01}.
}
\usage{
data(hisclass)
}
\description{
Characteristict of hisclass classes.
}

