% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hisclass_5}
\alias{hisclass_5}
\title{HISCLASS 5 meta data}
\format{A data frame with 6 rows and 4 variables:
\describe{
  \item{hisclass_5}{HISCLASS 5 code}
  \item{hisclass_relation}{Relation to HISCLASS}
  \item{hisclass_5_label}{Class label}
  \item{hisclass_5_description}{Desciription of class}
}}
\source{
Leeuwen, M.H.D.V. & Maas, I. (2011). \emph{HISCLASS: a
  historical international social class scheme. Leuven:
  Leuven University Press};  Mandemakers etal. (2013)
  \emph{HSN standardized, HISCO-coded and classified occupational
  titles, release 2013.01}.
}
\usage{
data(hisclass_5)
}
\description{
Characteristict of HISCLASS 5 classes.
}

