% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hisco}
\alias{hisco}
\title{Hisco codes}
\format{A data frame with 4123 rows and 9 variables:
\describe{
  \item{id}{id code}
  \item{original}{Original text}
  \item{standard}{Standardized text}
  \item{hisco}{HISCO basic code}
  \item{status}{HISCO status code}
  \item{relation}{HISCO relation code}
  \item{product}{HISCO  product code}
  \item{hisclass}{HISCLASS code}
  \item{hisclass_5}{HISCLASS code comprised in 5 classes}
  \item{hiscam_u1}{Universal HISCAM code}
  \item{hiscam_nl}{HISCAM code}
  \item{socpo}{SOCPO code}
  \item{hisclass_label}{HISCLASS label}
  \item{hisclass_5_label}{HISCLASS 5 label}
  \item{socpo_label}{SOCPO label}
  \item{release}{HISCO release version}
}}
\source{
The dataset is based on the Kees Mandemakers, Sanne Muurling,
  Ineke Maas, Bart Van de Putte, Richard L. Zijdeman, Paul Lambert,
  Marco H.D. van Leeuwen, Frans van Poppel and Andrew Miles,
  \emph{HSN standardized, HISCO-coded and classified occupational titles,
  release 2013.01} (IISG Amsterdam 2013); Marco H. D. van Leeuwen,
  Ineke Maas and Andrew Miles, \emph{HISCO. Historical International
  Standard Classification of Occupations} (Leuven University Press 2002)
  downloaded from \url{http://www.iisg.nl/hsn/data/occupations.html}
  \emph{Accessed, 2014-09-02}; Pre-processed by Glenn Sandström,
  Umeå university.
}
\usage{
data(hisco)
}
\description{
Hisco coding table, to historical classsystems,
  HISCLASS, HISCLASS 5, SOCPO and Universal HISCAM
}

