% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/hisco_to_ses.R
\name{hisco_to_ses}
\alias{hisco_to_ses}
\title{Hisco to SES classification}
\usage{
hisco_to_ses(x, ses = c("hisclass", "hisclass_5", "socpo", "hiscam_u1",
  "all"), status = NULL, relation = NULL, product = NULL, label = FALSE,
  name = "hisco", reference = NULL, messages = FALSE)
}
\arguments{
\item{x}{a vector of HISCO codes, a path to a csv file or data.frame that holds HISCO and
optional STATUS, RELATION and PRODUCT mappigns}

\item{ses}{ouput SES system}

\item{status}{name of STATUS variable or vector of STATUS codes}

\item{relation}{name of RELATION variable or vector of RELATION codes}

\item{product}{name of PRODUCT variable or vector of PRODUCT codes}

\item{label}{logical. TRUE will display label for SES such as "Forman".
FALSE will display the SES as a numeric code.}

\item{name}{character. Used when input is a data.frame to specify the
name of the column containing hisco codes.}

\item{reference}{character. Path to a csv file containing a reference table.}

\item{messages}{Print summary statistics on classification result.}
}
\description{
Function to classify historical occupations coded in HICSO
  (Historical International Standard Classification of Occupations)
  to HISCLASS 12, HISCLASS 5, SOCPO or HISCAM social class systems.
}
\author{
Johan Junkka \email{johan.junkka@gmail.com}
}

