% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{socpo}
\alias{socpo}
\title{SOCPO meta data}
\format{A data frame with 14 rows and 6 variables:
\describe{
  \item{socpo}{SOCPO code}
  \item{socpo_label}{Class name}
  \item{socpo_description}{Desciription of class}
}}
\source{
Leeuwen, M.H.D.V. & Maas, I. (2011). \emph{HISCLASS: a
  historical international social class scheme. Leuven:
  Leuven University Press};  Mandemakers etal. (2013)
  \emph{HSN standardized, HISCO-coded and classified occupational
  titles, release 2013.01}.NULL
}
\usage{
data(socpo)
}
\description{
Characteristict of SOCPO classes.
}

