% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/reference.R
\name{write_reference}
\alias{write_reference}
\title{Write hisco reference file to csv}
\usage{
write_reference(path)
}
\arguments{
\item{path}{character sting with name of file}
}
\description{
To manually update the hisco classifications you can
  start with the package reference file as a template.
  This function writes the reference file as csv file to
  a specified path.
}
\examples{
write_reference("my_hisco.csv")
unlink("my_hisco.csv")
}
\author{
Johan Junkka \email{johan.junkka@gmail.com}
}

